requirements.txt
openai
datamodel-code-generator
pyyaml
langchain